+++ hutool-core/src/main/java/cn/hutool/core/date/DateUtil.java
@@ -407,7 +407,7 @@ public class DateUtil {
 	}
 
 	/**
-	 * @return 当前日期所在年份的第几周
+	 * @return 当前日期所在月份的第几周
 	 */
 	public static int thisWeekOfMonth() {
 		return weekOfMonth(date());
@@ -1711,6 +1711,16 @@ public class DateUtil {
 	}
 
 	/**
+	 * 是否闰年
+	 *
+	 * @param year 年
+	 * @return 是否闰年
+	 */
+	public static boolean isLeapYear(int year) {
+		return new GregorianCalendar().isLeapYear(year);
+	}
+
+	/**
 	 * 计算相对于dateToCompare的年龄，长用于计算指定生日在某年的年龄
 	 *
 	 * @param birthDay      生日
@@ -1725,17 +1735,20 @@ public class DateUtil {
 			throw new IllegalArgumentException(StrUtil.format("Birthday is after date {}!", formatDate(dateToCompare)));
 		}
 
-		int year = cal.get(Calendar.YEAR);
-		int month = cal.get(Calendar.MONTH);
-		int dayOfMonth = cal.get(Calendar.DAY_OF_MONTH);
+		final int year = cal.get(Calendar.YEAR);
+		final int month = cal.get(Calendar.MONTH);
+		final int dayOfMonth = cal.get(Calendar.DAY_OF_MONTH);
+		final boolean isLastDayOfMonth = dayOfMonth == cal.getActualMaximum(Calendar.DAY_OF_MONTH);
 
 		cal.setTime(birthDay);
 		int age = year - cal.get(Calendar.YEAR);
 
-		int monthBirth = cal.get(Calendar.MONTH);
+		final int monthBirth = cal.get(Calendar.MONTH);
 		if (month == monthBirth) {
-			int dayOfMonthBirth = cal.get(Calendar.DAY_OF_MONTH);
-			if (dayOfMonth < dayOfMonthBirth) {
+
+			final int dayOfMonthBirth = cal.get(Calendar.DAY_OF_MONTH);
+			final boolean isLastDayOfMonthBirth = dayOfMonthBirth == cal.getActualMaximum(Calendar.DAY_OF_MONTH);
+			if ((false == isLastDayOfMonth || false == isLastDayOfMonthBirth) && dayOfMonth < dayOfMonthBirth) {
 				// 如果生日在当月，但是未达到生日当天的日期，年龄减一
 				age--;
 			}
@@ -1745,16 +1758,6 @@ public class DateUtil {
 		}
 
 		return age;
-	}
-
-	/**
-	 * 是否闰年
-	 *
-	 * @param year 年
-	 * @return 是否闰年
-	 */
-	public static boolean isLeapYear(int year) {
-		return new GregorianCalendar().isLeapYear(year);
 	}
 
 	/**
@@ -1967,6 +1970,17 @@ public class DateUtil {
 	}
 
 	/**
+	 * Calendar{@link Instant}对象
+	 *
+	 * @param calendar Date对象
+	 * @return {@link Instant}对象
+	 * @since 5.0.5
+	 */
+	public static Instant toInstant(Calendar calendar) {
+		return null == calendar ? null : calendar.toInstant();
+	}
+
+	/**
 	 * Date对象转换为{@link Instant}对象
 	 *
 	 * @param temporalAccessor Date对象
@@ -2001,8 +2015,42 @@ public class DateUtil {
 
 		return result;
 	}
-	// ------------------------------------------------------------------------ Private method start
 
+	/**
+	 * {@link Instant} 转换为 {@link LocalDateTime}，使用系统默认时区
+	 *
+	 * @param instant {@link Instant}
+	 * @return {@link LocalDateTime}
+	 * @since 5.0.5
+	 */
+	public static LocalDateTime toLocalDateTime(Instant instant){
+		return LocalDateTime.ofInstant(instant, ZoneId.systemDefault());
+	}
+
+	/**
+	 * {@link Calendar} 转换为 {@link LocalDateTime}，使用系统默认时区
+	 *
+	 * @param calendar {@link Calendar}
+	 * @return {@link LocalDateTime}
+	 * @since 5.0.5
+	 */
+	public static LocalDateTime toLocalDateTime(Calendar calendar){
+		return LocalDateTime.ofInstant(calendar.toInstant(), calendar.getTimeZone().toZoneId());
+	}
+
+	/**
+	 * {@link Date} 转换为 {@link LocalDateTime}，使用系统默认时区
+	 *
+	 * @param date {@link Calendar}
+	 * @return {@link LocalDateTime}
+	 * @since 5.0.5
+	 */
+	public static LocalDateTime toLocalDateTime(Date date){
+		final DateTime dateTime = date(date);
+		return LocalDateTime.ofInstant(dateTime.toInstant(), dateTime.getZoneId());
+	}
+
+	// ------------------------------------------------------------------------ Private method start
 	/**
 	 * 获得指定日期年份和季节<br>
 	 * 格式：[20131]表示2013年第一季度