+++ hutool-core/src/test/java/cn/hutool/core/date/DateBetweenTest.java
@@ -1,11 +1,10 @@
 package cn.hutool.core.date;
 
-import java.util.Date;
-
+import cn.hutool.core.date.BetweenFormater.Level;
 import org.junit.Assert;
 import org.junit.Test;
 
-import cn.hutool.core.date.BetweenFormater.Level;
+import java.util.Date;
 
 public class DateBetweenTest {
 
@@ -26,6 +25,14 @@ public class DateBetweenTest {
 		Date end2 = DateUtil.parse("2018-02-01 11:23:46");
 		long betweenYear2 = new DateBetween(start2, end2).betweenYear(false);
 		Assert.assertEquals(0, betweenYear2);
+	}
+
+	@Test
+	public void betweenYearTest2() {
+		Date start = DateUtil.parse("2000-02-29");
+		Date end = DateUtil.parse("2018-02-28");
+		long betweenYear = new DateBetween(start, end).betweenYear(false);
+		Assert.assertEquals(18, betweenYear);
 	}
 
 	@Test